pyproject.toml
[project]
name = "fastapi-vs-django-experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.15",
    "daphne>=4.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "typer>=0.12",
    "httpx>=0.27",
    "psutil>=6.0",
]

[project.scripts]
bench = "eval.cli:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["eval*"]

[dependency-groups]
dev = ["pre-commit>=4.0"]
dashboard = ["streamlit>=1.40", "plotly>=5.24", "pandas>=2.2"]

[tool.uv]
package = true
